requirements.txt
django>=5.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
mysqlclient>=2.2.0
celery>=5.3.0
drf-yasg>=1.21.0
pika>=1.3.0